RALPH: Review refinements for issue #30 (no functional change)
- freeplast-woo: drop the redundant (int) cast on fpw_cart_line_count() in
  the fragment (the function declares ': int').
- basket-count.js: name the repeated 'wc/store/cart' store as CART_STORE
  (matching the existing SELECTOR constant), hoist String(count) out of the
  render loop, and read window.wp.data once in lineCount().
- test-woo-adapter.php: name the resolver filter's variables (,
  ) and split the dense one-liner for readability.

Offline checks unchanged: 30 local assertions + 14 syntax/dependency/
deployment checks pass.

+++ wordpress/wp-content/plugins/freeplast-woo/freeplast-woo.php
@@ -67,7 +67,7 @@ function fpw_cart_line_count(): int {
 
 /** The header count participates in Woo's native add-to-cart fragment refresh. */
 add_filter( 'woocommerce_add_to_cart_fragments', static function ( $fragments ) {
-	$fragments['span.fpw-basket-count'] = '<span class="fpw-basket-count">' . (int) fpw_cart_line_count() . '</span>';
+	$fragments['span.fpw-basket-count'] = '<span class="fpw-basket-count">' . fpw_cart_line_count() . '</span>';
 	return $fragments;
 } );
